+++ Factory/SwaggerResolverFactory.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
 namespace Linkin\Bundle\SwaggerResolverBundle\Factory;
 
 use Linkin\Bundle\SwaggerResolverBundle\Builder\SwaggerResolverBuilder;
-use Linkin\Bundle\SwaggerResolverBundle\Configuration\SwaggerConfiguration;
+use Linkin\Bundle\SwaggerResolverBundle\Configuration\SwaggerConfigurationInterface;
 use Linkin\Bundle\SwaggerResolverBundle\Resolver\SwaggerResolver;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\RouterInterface;
@@ -38,18 +38,18 @@ class SwaggerResolverFactory
     private $router;
 
     /**
-     * @var SwaggerConfiguration
+     * @var SwaggerConfigurationInterface
      */
     private $swaggerConfiguration;
 
     /**
      * @param SwaggerResolverBuilder $builder
-     * @param SwaggerConfiguration $configuration
+     * @param SwaggerConfigurationInterface $configuration
      * @param RouterInterface $router
      */
     public function __construct(
         SwaggerResolverBuilder $builder,
-        SwaggerConfiguration $configuration,
+        SwaggerConfigurationInterface $configuration,
         RouterInterface $router
     ) {
         $this->builder = $builder;